Fix Hibernate schema-validation mismatches against remote MySQL
- Map tinyint columns as Byte instead of Integer (Branch.estimatedWaitMinutes,
  OrderItem.spoonCount/dryIceMinutes, OrderItemFlavor.selectionOrder/quantity,
  Product.selectableFlavorCount)
- Set hibernate.type.preferred_boolean_jdbc_type=TINYINT so Boolean fields
  match MySQL boolean-as-tinyint storage
- Add explicit length=65535 to @Lob String fields so Hibernate maps them to
  TEXT instead of TINYTEXT (ChatMessage, Event, Notice)
- Point default DB_HOST/DB_PASSWORD at shared dev server
- Specify mainClass in spring-boot-maven-plugin config

+++ backend/src/main/java/com/kiosk/domain/order/OrderItem.java
@@ -61,10 +61,10 @@ public class OrderItem {
 
     @Column(name = "spoon_count", nullable = false)
     @Builder.Default
-    private Integer spoonCount = 0;
+    private Byte spoonCount = 0;
 
     @Column(name = "dry_ice_minutes")
-    private Integer dryIceMinutes;
+    private Byte dryIceMinutes;
 
     @Column(name = "request_note", length = 500)
     private String requestNote;
